requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.20.1
Pillow==8.1.0
